+++ CHANGES.md
@@ -7,6 +7,7 @@ dev
 - Install bzip2 and rsync on OpenSUSE distros.
 - Add a `Dockerfile_opam.deprecated` container for being able to turn off older distros. 
 - Install `which` into OpenSUSE containers by default.
+- Use `+trunk` suffix for dev versions of compiler.
 
 v6.0.0 2018-11-15 Cambridge
 ---------------------------